+++ python/sdss_access/path/path.py
@@ -66,7 +66,8 @@ class BasePath(object):
 
     def replant_tree(self):
         ''' replants the tree based on release '''
-        tree.replant_tree(self.release)
+        release = self.release.lower() if self.release else None
+        tree.replant_tree(release)
 
     def _input_templates(self):
         """Read the path template file.