+++ python/sdss_access/tests/conftest.py
@@ -11,101 +11,24 @@
 from __future__ import print_function, division, absolute_import
 import os
 import pytest
+import yaml
 
-from sdss_access import RsyncAccess
-from sdss_access import CurlAccess
+from sdss_access import RsyncAccess, HttpAccess, CurlAccess
 from sdss_access.path import Path
 
 
-releases = ['MPL-5']
-plateifus = ['8485-1901']
-mpldict = {'MPL-6': ('v2_3_1', '2.1.3'), 'MPL-5': ('v2_0_1', '2.0.2')}
-surveys = ['manga']
+# PYTEST MODIFIERS
+# -----------------
+def pytest_addoption(parser):
+    """Add new options"""
+    # run slow tests
+    parser.addoption('--runslow', action='store_true', default=False, help='Run slow tests.')
 
 
-@pytest.fixture(scope='session', params=releases)
-def get_release(request):
-    ''' MaNGA release fixture '''
-    return request.param
-
-
-@pytest.fixture(scope='session', params=plateifus)
-def get_plateifu(request):
-    ''' MaNGA plateifu fixture '''
-    return request.param
-
-
-class Survey(object):
-    ''' Survey class '''
-    def __init__(self, survey):
-        self.survey = survey
-        if self.survey == 'manga':
-            self.setup_manga()
-
-    def setup_manga(self):
-        self.path_kwargs = self.get_path_kwargs()
-
-    def get_path_kwargs(self):
-        if self.survey == 'manga':
-            rkwargs = {'plate': None, 'ifu': None, 'drpver': None, 'dapver': None, 'dir3d': None,
-                       'mpl': None, 'bintype': '*', 'n': '**', 'mode': '*', 'daptype': '*'}
-        return rkwargs
-
-    def set_data(self, release):
-        if self.survey == 'manga':
-            self.release = release
-            self.names = ['mangacube']
-            if '4' in release:
-                self.names.extend(['mangadefault', 'mangamap'])
-            else:
-                self.names.extend(['mangadap5'])
-
-
-@pytest.fixture(scope='session', params=surveys)
-def survey(request):
-    ''' fixture to generate a survey'''
-    survey = Survey(request.param)
-    yield survey
-    survey = None
-
-
-@pytest.fixture(scope='session')
-def init_manga_survey(survey, get_release, get_plateifu):
-    ''' create different surveys with different parameters '''
-
-    drpver, dapver = mpldict[get_release]
-    plate, ifu = get_plateifu.split('-')
-    survey.path_kwargs['drpver'] = drpver
-    survey.path_kwargs['dapver'] = dapver
-    survey.path_kwargs['plate'] = plate
-    survey.path_kwargs['ifu'] = ifu
-    survey.set_data(get_release)
-    yield survey
-
-
-@pytest.fixture(scope='module', params=['mangadap5', 'mangacube'])
-def data(request, init_manga_survey):
-    ''' fixture to generate data '''
-    fillkwargs = {'plate': init_manga_survey.path_kwargs['plate'], 'ifu': init_manga_survey.path_kwargs['ifu'],
-                  'drpver': init_manga_survey.path_kwargs['drpver'], 'dapver': init_manga_survey.path_kwargs['dapver']}
-    data_dict = {'mangadefault': {'files': 'mangadap-{plate}-{ifu}-default.fits.gz'.format(**fillkwargs),
-                                  'loc': 'mangawork/manga/spectro/analysis/{drpver}/{dapver}/default/{plate}/'.format(**fillkwargs),
-                                  'single': 'mangadap-{plate}-{ifu}-default.fits.gz'.format(**fillkwargs),
-                                  'count': 1},
-                 'mangamap': {'files': 'manga-{plate}-{ifu}-LOGCUBE_MAPS-*-0**.fits.gz'.format(**fillkwargs),
-                              'loc': 'mangawork/manga/spectro/analysis/{drpver}/{dapver}/full/{plate}/{ifu}/'.format(**fillkwargs),
-                              'single': 'manga-{plate}-{ifu}-LOGCUBE_MAPS-NONE-003.fits.gz'.format(**fillkwargs),
-                              'count': 21},
-                 'mangadap5': {'files': 'manga-{plate}-{ifu}-*-SPX-GAU-MILESHC.fits.gz'.format(**fillkwargs),
-                               'loc': 'mangawork/manga/spectro/analysis/{drpver}/{dapver}/*/{plate}/{ifu}/'.format(**fillkwargs),
-                               'single': 'manga-{plate}-{ifu}-MAPS-SPX-GAU-MILESHC.fits.gz'.format(**fillkwargs),
-                               'count': 2},
-                 'mangacube': {'files': 'manga-{plate}-{ifu}-LOGCUBE.fits.gz'.format(**fillkwargs),
-                               'loc': 'mangawork/manga/spectro/redux/{drpver}/{plate}/stack/'.format(**fillkwargs),
-                               'single': 'manga-{plate}-{ifu}-LOGCUBE.fits.gz'.format(**fillkwargs),
-                               'count': 1}
-                 }
-    return (request.param, data_dict[request.param])
+def pytest_runtest_setup(item):
+    """Skip slow tests."""
+    if 'slow' in item.keywords and not item.config.getoption('--runslow'):
+        pytest.skip('Requires --runslow option to run.')
 
 
 @pytest.fixture()
@@ -116,64 +39,140 @@ def path():
     path = None
 
 
-@pytest.fixture(scope='function')
-def rsync():
-    ''' fixture to create generic rsync object '''
-    rsync = RsyncAccess(label='test_rsync')
+# releases to parametrize over
+releases = ['work', 'DR15']
+
+
+# read in test data parameters and also get paths
+def get_data():
+    ''' Retrieves the test data from the paths.yaml file '''
+
+    with open(os.path.join(os.path.dirname(__file__), 'data/paths.yaml')) as f:
+        data = yaml.load(f, Loader=yaml.SafeLoader)
+    return data
+
+data = get_data()
+paths = data.get('paths')
+
+
+@pytest.fixture(scope='session', params=releases)
+def release(request):
+    ''' release fixture '''
+    return request.param
+
+
+@pytest.fixture(scope='session', params=paths)
+def datapath(request):
+    ''' parametrizes over the paths in test data'''
+    return request.param
+
+
+@pytest.fixture(scope='session')
+def expdata(release, datapath):
+    ''' fixture to yield expected source data based on test data '''
+    remote = data.get('remote_base')
+    # remote base
+    base = remote['work'] if release == 'work' else remote['public']
+    # work or DR directory
+    name = datapath['work'] if release == 'work' else ''
+    # file location
+    location = os.path.join(name, datapath['location'])
+    # full source file location
+    source = os.path.join(base, 'sas' if release == 'work' else release.lower(), location)
+    # full final file location
+    destination = os.path.join(os.getenv('SAS_BASE_DIR'), '' if release == 'work' else release.lower(), location)
+    # combined dict
+    result = {'name': datapath['name'], 'params': datapath['params'], 'base': base,
+              'location': location, 'source': source, 'destination': destination, 'release': release.lower()}
+    yield result
+    result = None
+
+
+@pytest.fixture(scope='session')
+def inittask(expdata):
+    ''' fixture to yield expected initial stream task based on test data '''
+
+    patch = '' if expdata['release'] == 'work' else expdata['release']
+    loc = os.path.join(patch, expdata['location'])
+    task = [{'location': loc, 'source': expdata['source'],
+             'destination': expdata['destination'], 'exists': None}]
+    yield task
+    task = None
+
+
+@pytest.fixture(scope='session')
+def finaltask(expdata):
+    ''' fixture to yield expected final stream task based on test data '''
+
+    task = [{'location': expdata['location'], 'source': expdata['source'],
+             'destination': expdata['destination'], 'exists': None}]
+    yield task
+    task = None
+
+
+@pytest.fixture(scope='session')
+def rsync(release):
+    ''' fixture to create generic rsync object - parametrized by release '''
+
+    if 'DR' in release:
+        rsync = RsyncAccess(label='test_rsync', public=True, release=release)
+    else:
+        rsync = RsyncAccess(label='test_rsync')
     rsync.remote()
     yield rsync
-
     # teardown
     rsync.reset()
     rsync = None
 
 
-@pytest.fixture(scope='function')
-def rsync_add(rsync, data, init_manga_survey):
-    ''' fixture to add data to an rsync object '''
-    name, paths = data
-    rsync.add(name, **init_manga_survey.path_kwargs)
-    rsync.location = paths['loc']
+@pytest.fixture(scope='session')
+def radd(rsync, expdata):
+    ''' fixture to add a path to an rsync object '''
+    rsync.add(expdata['name'], **expdata['params'])
     yield rsync
 
 
-@pytest.fixture(scope='function')
-def rsync_set(rsync_add, data):
-    ''' fixture to set the stream of an rsync object '''
-    name, paths = data
-    rsync_add.location = os.path.join(paths['loc'].replace('*', 'SPX-GAU-MILESHC'), paths['single'])
-    rsync_add.count = paths['count']
-    rsync_add.set_stream()
-    yield rsync_add
+@pytest.fixture(scope='session')
+def rstream(radd):
+    ''' fixture to set the stream for an parametrized rsync object '''
+    radd.set_stream()
+    yield radd
 
-@pytest.fixture(scope='function')
-def curl():
-    ''' fixture to create generic curl object '''
-    curl = CurlAccess(label='test_curl')
+
+@pytest.fixture(scope='session')
+def http(release):
+    if 'DR' in release:
+        http = HttpAccess(public=True, release=release)
+    else:
+        http = HttpAccess()
+    yield http
+    http = None
+
+
+@pytest.fixture(scope='session')
+def curl(release):
+    ''' fixture to create generic curl object - parametrized by release '''
+
+    if 'DR' in release:
+        curl = CurlAccess(label='test_curl', public=True, release=release)
+    else:
+        curl = CurlAccess(label='test_curl')
     curl.remote()
     yield curl
-
     # teardown
     curl.reset()
     curl = None
 
 
-@pytest.fixture(scope='function')
-def curl_add(curl, data, init_manga_survey):
-    ''' fixture to add data to an curl object '''
-    name, paths = data
-    curl.add(name, **init_manga_survey.path_kwargs)
-    curl.location = paths['loc']
+@pytest.fixture(scope='session')
+def cadd(curl, expdata):
+    ''' fixture to add a path to an curl object '''
+    curl.add(expdata['name'], **expdata['params'])
     yield curl
 
 
-@pytest.fixture(scope='function')
-def curl_set(curl_add, data):
-    ''' fixture to set the stream of an curl object '''
-    name, paths = data
-    curl_add.location = os.path.join(paths['loc'].replace('*', 'SPX-GAU-MILESHC'), paths['single'])
-    curl_add.count = paths['count']
-    curl_add.set_stream()
-    yield curl_add
-
-
+@pytest.fixture(scope='session')
+def cstream(cadd):
+    ''' fixture to set the stream for an parametrized curl object '''
+    cadd.set_stream()
+    yield cadd